+++ src/plasma/test/serialization_tests.c
@@ -150,11 +150,11 @@ TEST plasma_seal_request_test(void) {
 TEST plasma_seal_reply_test(void) {
   int fd = create_temp_file();
   object_id object_id1 = globally_unique_id();
-  uint8_t error1 = 5;
+  int error1 = 5;
   plasma_send_seal_reply(fd, object_id1, error1);
   uint8_t *data = read_message_from_file(fd, MessageType_PlasmaSealReply);
   object_id object_id2;
-  uint8_t error2;
+  int error2;
   plasma_read_seal_reply(data, &object_id2, &error2);
   ASSERT(object_ids_equal(object_id1, object_id2));
   ASSERT(error1 == error2);
